Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-createOffer-expected.txt


PASS createOffer() with no argument from newly created RTCPeerConnection should succeed 
PASS createOffer() and then setLocalDescription() should succeed 
PASS createOffer() after connection is closed should reject with InvalidStateError 
FAIL When media stream is added when createOffer() is running in parallel, the result offer should contain the new media stream assert_equals: Expect m=audio line to be found in offer SDP expected 1 but got 0
